+++ spring-cloud-starter-tencent-polaris-discovery/src/main/java/com/tencent/cloud/polaris/registry/PolarisRegistration.java
@@ -19,7 +19,6 @@
 package com.tencent.cloud.polaris.registry;
 
 import java.net.URI;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -54,7 +53,7 @@ public class PolarisRegistration implements Registration {
 
 	private Map<String, String> metadata;
 
-	private final String host;
+	private String host;
 
 	public PolarisRegistration(
 			PolarisDiscoveryProperties polarisDiscoveryProperties,
@@ -76,6 +75,10 @@ public class PolarisRegistration implements Registration {
 	@Override
 	public String getHost() {
 		return host;
+	}
+
+	public void setHost(String host) {
+		this.host = host;
 	}
 
 	@Override
@@ -108,7 +111,7 @@ public class PolarisRegistration implements Registration {
 
 			instanceMetadata.putAll(staticMetadataManager.getMergedStaticMetadata());
 
-			this.metadata = Collections.unmodifiableMap(instanceMetadata);
+			this.metadata = instanceMetadata;
 		}
 		return metadata;
 	}