+++ spring-cloud-starter-tencent-polaris-router/src/main/java/com/tencent/cloud/polaris/router/config/RouterAutoConfiguration.java
@@ -45,6 +45,7 @@ import static org.springframework.core.Ordered.HIGHEST_PRECEDENCE;
  * @author lepdou 2022-05-11
  */
 @Configuration(proxyBeanMethods = false)
+@ConditionalOnPolarisRouterEnabled
 @LoadBalancerClients(defaultConfiguration = LoadBalancerConfiguration.class)
 @Import({PolarisNearByRouterProperties.class, PolarisMetadataRouterProperties.class, PolarisRuleBasedRouterProperties.class})
 public class RouterAutoConfiguration {