+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 ---
 
 - [Optimize: remove discovery module useless code](https://github.com/Tencent/spring-cloud-tencent/pull/596)
+- [Optimize: optimize configuration conditional & optimize config data tips"](https://github.com/Tencent/spring-cloud-tencent/pull/604)